dismiss keyguard for CallNotificationActivity (WIP, see TODO)
Whenever i click the button to accept the call, CallActivity is started by intent.
CallActivity itself has the same logic to dismiss the keyguard.

The problem is: When switching from CallNotificationActivity to CallActivity the lockscreen is shown indeed!
So i guess Android recognizes "oh, CallNotificationActivity is gone. Let's show the lockscreen" before it recognizes that
the new activity also dismiss the lockscreen ?!

I fear for this scenario it was not the best idea to have 2 different activies.. (before it was only 1 activity and 2 condcutor controllers..)

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>