Show profile action on avatar click
For every click on a avatar in a group chat it will be tried to receive a
hover card from the Nextcloud server. The endpoint returns multiple
actions. Those actions will be filtered for the app ids 'spreed',
'email' and 'profile'. Other will be ignored.
The received filtered actions will be shown in a bottom sheet.

Resolves: #1644

Signed-off-by: Tim Krüger <t@timkrueger.me>